requirements.txt
# Core dependencies for PDF RAG system
streamlit>=1.25.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
pymupdf>=1.22.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.21.0
pandas>=1.5.0
requests>=2.28.0
scikit-learn>=1.2.0

# Optional for better compatibility
python-dotenv>=0.19.0
regex>=2022.0.0